src/main/java/Student.java


public class Student {

  private String firstName;
  private String lastName;
  private String email;

  public Student(String firstName, String lastName, String email) {
    this.firstName = firstName;
    this.lastName = lastName;
    this.email = email;
  }

  public String getFirstName() {
    return this.firstName;
  }

  public String getLastName() {
    return this.lastName;
  }

  public String getEmail() {
    return this.email;
  }

  public String getDirectoryEntry() {
    return this.lastName + ", " + this.firstName + ", " + this.email;
  }
}
